+++ src/test/scala/com/dataintuitive/viash/MainTestDockerSuite.scala
@@ -17,6 +17,7 @@ class MainTestDockerSuite extends FunSuite with BeforeAndAfterAll {
   private val configFailedBuildFile = getClass.getResource("/testbash/config_failed_build.vsh.yaml").getPath
   private val configNonexistentTestFile = getClass.getResource("/testbash/config_nonexistent_test.vsh.yaml").getPath
   private val configWithSpacesFile = getClass.getResource("/testbash/config test.vsh.yaml").getPath.replaceAll("%20", " ")
+  private val configLegacyTestFile = getClass.getResource("/testbash/config_legacy.vsh.yaml").getPath
 
   private val configMissingFunctionalityFile = getClass.getResource("/testbash/invalid_configs/config_missing_functionality.vsh.yaml").getPath
   private val configTextFile = getClass.getResource("/testbash/invalid_configs/config.txt").getPath
@@ -130,6 +131,20 @@ class MainTestDockerSuite extends FunSuite with BeforeAndAfterAll {
       "test",
       "-p", "docker",
       configWithSpacesFile
+    )
+
+    assert(testText.contains("Running tests in temporary directory: "))
+    assert(testText.contains("SUCCESS! All 2 out of 2 test scripts succeeded!"))
+    assert(testText.contains("Cleaning up temporary directory"))
+
+    checkTempDirAndRemove(testText, false)
+  }
+
+    test("Check standard test output with legacy 'tests' definition", DockerTest) {
+    val testText = TestHelper.testMain(
+      "test",
+      "-p", "docker",
+      configLegacyTestFile
     )
 
     assert(testText.contains("Running tests in temporary directory: "))